renaming pollConfiguration to Election service and controller
