+++ src/main/java/org/sysc4907/votingsystem/Elections/ElectionController.java
@@ -1,4 +1,4 @@
-package org.sysc4907.votingsystem.PollConfiguration;
+package org.sysc4907.votingsystem.Elections;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.format.annotation.DateTimeFormat;
@@ -8,7 +8,6 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.multipart.MultipartFile;
-import org.sysc4907.votingsystem.Elections.Election;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
@@ -16,18 +15,18 @@ import java.util.Arrays;
 import java.util.List;
 
 @Controller
-public class PollConfigurationController {
+public class ElectionController {
 
-    private final PollConfigurationService pollConfigurationService;
+    private final ElectionService electionService;
 
     @Autowired
-    public PollConfigurationController(PollConfigurationService pollConfigurationService) {
-        this.pollConfigurationService = pollConfigurationService;
+    public ElectionController(ElectionService electionService) {
+        this.electionService = electionService;
     }
 
     @GetMapping
     public Election getElectionDetails() {
-        return pollConfigurationService.getElection();
+        return electionService.getElection();
     }
 
     @GetMapping("/create-election")
@@ -53,7 +52,7 @@ public class PollConfigurationController {
         List<String> splitCandidates = Arrays.asList(candidates.split("\\r?\\n"));
         model.addAttribute("candidates", splitCandidates);
 
-        if (pollConfigurationService.validateAndConfigurePoll(
+        if (electionService.validateAndConfigurePoll(
                 startDate, startTime, endDate, endTime, name, candidates, voterKeys) == true) {
             return "successful-poll-config";
         }
